requirements.txt
fastapi
uvicorn[standard]
python-multipart
docling
chromadb
sentence-transformers
groq
python-dotenv
pydantic
